cmapss_simulator/dashboard/requirements.txt
streamlit==1.29.0
pandas==2.1.3
google-cloud-bigquery==3.13.0
plotly==5.18.0
db-dtypes==1.1.1
